git-svn-id: svn://192.168.1.10/public/CoffeeMud@9661 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Abilities/Common/Merchant.java
@@ -41,7 +41,7 @@ public class Merchant extends CommonSkill implements ShopKeeper
 	public String name(){ return "Marketeering";}
 	private static final String[] triggerStrings = {"MARKET"};
 	public String[] triggerStrings(){return triggerStrings;}
-	public int overrideMana(){return 5;}
+	public short overrideMana(){return 5;}
 	public boolean isAutoInvoked(){return true;}
 	public boolean canBeUninvoked(){return false;}
 	protected ExpertiseLibrary.SkillCostDefinition getRawTrainingCost() { return CMProps.getSkillTrainCostFormula(ID()); }
