+++ com/planet_ink/coffee_mud/Abilities/Druid/Chant_SpeedTime.java
@@ -44,7 +44,7 @@ public class Chant_SpeedTime extends Chant
 	public int abstractQuality(){ return Ability.QUALITY_INDIFFERENT;}
 	protected int canAffectCode(){return 0;}
 	protected int canTargetCode(){return 0;}
-	protected int overrideMana(){return 100;}
+	protected short overrideMana(){return 100;}
 
 	public boolean invoke(MOB mob, Vector commands, Physical givenTarget, boolean auto, int asLevel)
 	{
