+++ com/planet_ink/coffee_mud/Abilities/Druid/Chant_VineWeave.java
@@ -41,7 +41,7 @@ public class Chant_VineWeave extends Chant
 	public int abstractQuality(){return Ability.QUALITY_INDIFFERENT;}
 	protected int canAffectCode(){return 0;}
 	protected int canTargetCode(){return 0;}
-	protected int overrideMana(){return 50;}
+	protected short overrideMana(){return 50;}
 	public int classificationCode(){return Ability.ACODE_CHANT|Ability.DOMAIN_PLANTGROWTH;}
 
 	public boolean invoke(MOB mob, Vector commands, Physical givenTarget, boolean auto, int asLevel)
