+++ com/planet_ink/coffee_mud/Abilities/Prayers/Prayer_BloodHearth.java
@@ -44,7 +44,7 @@ public class Prayer_BloodHearth extends Prayer
 	public int abstractQuality(){ return Ability.QUALITY_MALICIOUS;}
 	protected int canAffectCode(){return CAN_ROOMS;}
 	protected int canTargetCode(){return CAN_ROOMS;}
-	protected int overrideMana(){return Integer.MAX_VALUE;}
+	protected short overrideMana(){return Short.MAX_VALUE;}
 	public long flags(){return Ability.FLAG_UNHOLY;}
 
 	public boolean okMessage(final Environmental myHost, final CMMsg msg)
