+++ com/planet_ink/coffee_mud/Abilities/Spells/Spell_AnimateWeapon.java
@@ -40,7 +40,7 @@ public class Spell_AnimateWeapon extends Spell
 	public int abstractQuality(){ return Ability.QUALITY_INDIFFERENT;}
 	protected int canTargetCode(){return CAN_ITEMS;}
 	public int classificationCode(){	return Ability.ACODE_SPELL|Ability.DOMAIN_ENCHANTMENT;}
-	public int overrideMana(){return 100;}
+	public short overrideMana(){return 100;}
 
 	public boolean tick(Tickable ticking, int tickID)
 	{
