+++ com/planet_ink/coffee_mud/Abilities/Spells/Spell_BaseClanEq.java
@@ -45,7 +45,7 @@ public class Spell_BaseClanEq extends Spell
 	public int abstractQuality(){ return Ability.QUALITY_INDIFFERENT;}
 	public int classificationCode(){return Ability.ACODE_SPELL|Ability.DOMAIN_ENCHANTMENT;}
 	public long flags(){return super.flags()|Ability.FLAG_CLANMAGIC;}
-	protected int overrideMana(){return Integer.MAX_VALUE;}
+	protected short overrideMana(){return Short.MAX_VALUE;}
 	protected String type="";
 	protected boolean disregardsArmorCheck(MOB mob){return true;}
