+++ com/planet_ink/coffee_mud/Abilities/Spells/Spell_Gate.java
@@ -38,7 +38,7 @@ public class Spell_Gate extends Spell
 	public String name(){return "Gate";}
 	protected int canTargetCode(){return 0;}
 	public int classificationCode(){ return Ability.ACODE_SPELL|Ability.DOMAIN_CONJURATION;}
-	protected int overrideMana(){return Integer.MAX_VALUE-50;}
+	protected short overrideMana(){return Short.MAX_VALUE-50;}
 	public long flags(){return Ability.FLAG_TRANSPORTING;}
 	public int abstractQuality(){ return Ability.QUALITY_INDIFFERENT;}
