+++ com/planet_ink/coffee_mud/Abilities/Spells/Spell_TimeStop.java
@@ -40,7 +40,7 @@ public class Spell_TimeStop extends Spell
 	public String displayText(){return "(Time is Stopped)";}
 	protected int canAffectCode(){return CAN_ROOMS|CAN_MOBS;}
 	protected int canTargetCode(){return 0;}
-	protected int overrideMana(){return 100;}
+	protected short overrideMana(){return 100;}
 	public int classificationCode(){return Ability.ACODE_SPELL|Ability.DOMAIN_ALTERATION;}
 	public int abstractQuality(){ return Ability.QUALITY_MALICIOUS;}
