+++ com/planet_ink/coffee_mud/Abilities/Spells/Spell_WordRecall.java
@@ -40,7 +40,7 @@ public class Spell_WordRecall extends Spell
 	protected int canTargetCode(){return 0;}
 	public int abstractQuality(){return Ability.QUALITY_INDIFFERENT;}
 	public int classificationCode(){ return Ability.ACODE_SPELL|Ability.DOMAIN_CONJURATION;}
-	protected int overrideMana(){return Integer.MAX_VALUE-90;}
+	protected short overrideMana(){return Short.MAX_VALUE-90;}
 	public long flags(){return Ability.FLAG_TRANSPORTING;}
 
 	protected int verbalCastCode(MOB mob, Physical target, boolean auto)
