+++ com/planet_ink/coffee_mud/Abilities/Thief/Thief_Snipe.java
@@ -43,7 +43,7 @@ public class Thief_Snipe extends ThiefSkill
 	public int abstractQuality(){return Ability.QUALITY_OK_OTHERS;}
 	private static final String[] triggerStrings = {"SNIPE"};
 	public String[] triggerStrings(){return triggerStrings;}
-	protected int overrideMana(){return 100;}
+	protected short overrideMana(){return 100;}
 	public int usageType(){return USAGE_MOVEMENT|USAGE_MANA;}
 	public int classificationCode(){return Ability.ACODE_THIEF_SKILL|Ability.DOMAIN_DIRTYFIGHTING;}
